+++ src/main/groovy/org/fundacionjala/gradle/plugins/enforce/tasks/salesforce/deployment/Update.groovy
@@ -51,9 +51,8 @@ class Update extends Deployment {
         verifyParameter()
         excludeFilesFromFilesChanged()
         showFilesChanged()
-        if (packageGenerator.fileTrackerMap.isEmpty()) {
-            logger.quiet(Constants.NOT_FILES_CHANGED)
-            return
+        if(isEmptyChangedFiles()) {
+            return;
         }
         createDestructive()
         createPackage()
@@ -78,6 +77,7 @@ class Update extends Deployment {
                 filesToCopy.add(new File(Paths.get(projectPath, nameFile).toString()))
             }
         }
+        filesToCopy.sort()
         packageGenerator.buildPackage(updatePackagePath)
         combinePackageToUpdate(updatePackagePath)
     }
@@ -139,7 +139,10 @@ class Update extends Deployment {
     /**
      * Prints files changed
      */
-    def showFilesChanged() {
+    public void showFilesChanged() {
+        if (isEmptyChangedFiles()) {
+            logger.quiet(Constants.NOT_FILES_CHANGED)
+        }
         if (packageGenerator.fileTrackerMap.size() > 0) {
             logger.quiet("*********************************************")
             logger.quiet("              Status Files Changed             ")
@@ -149,6 +152,17 @@ class Update extends Deployment {
             }
             logger.quiet("*********************************************")
         }
+    }
+
+    /**
+     * Verifies if there are not changed files
+     * @return boolean
+     */
+    private boolean isEmptyChangedFiles() {
+        if(packageGenerator.fileTrackerMap.isEmpty()) {
+            return true
+        }
+        return false
     }
 
     /**
@@ -175,8 +189,10 @@ class Update extends Deployment {
     /**
      * ExcludeFiles from filesExcludes map
      */
-    private void excludeFilesFromFilesChanged() {
-        ArrayList<File> filesFiltered = excludeFiles(packageGenerator.getFiles())
-        filesExcludes = packageGenerator.excludeFiles(filesFiltered)
+    public void excludeFilesFromFilesChanged() {
+        ArrayList<File> files = packageGenerator.getFiles(projectPath)
+        ArrayList<File> filesFiltered = excludeFiles(files)
+        packageGenerator.updateFileTracker(filesFiltered)
+        filesExcludes = files - filesFiltered
     }
 }