+++ src/test/groovy/org/fundacionjala/gradle/plugins/enforce/tasks/salesforce/unittest/RunTestTaskTest.groovy
@@ -60,19 +60,52 @@ class RunTestTaskTest extends Specification {
 
     def "Should get the class names from a wildcard"() {
         given:
-        project.enforce {
-            srcPath = SRC_PATH
-            standardObjects = ["Q2w_Test__c.object"]
-            tool = "metadata"
-            poll = 200
-            waitTime = 10
-        }
-        def classNames = []
+            project.enforce {
+                srcPath = SRC_PATH
+                standardObjects = ["Q2w_Test__c.object"]
+                tool = "metadata"
+                poll = 200
+                waitTime = 10
+            }
+            def classNames = []
         when:
-        RunTestTask runTestTask = project.tasks.findByName(RUN_TEST_TASK_NAME) as RunTestTask
-        def path = Paths.get(SRC_PATH, "test").toString()
-        classNames = runTestTask.getClassNames(path, "*Test*")
+            RunTestTask runTestTask = project.tasks.findByName(RUN_TEST_TASK_NAME) as RunTestTask
+            def path = Paths.get(SRC_PATH, "test").toString()
+            classNames = runTestTask.getClassNames(path, "*Test*")
         then:
-        classNames.sort() == ["FGW_Console_CTRLTest", "FGW_APIFactoryTest"].sort()
+            classNames.sort() == ["FGW_Console_CTRLTest", "FGW_APIFactoryTest"].sort()
+    }
+
+    def "Test should return the class name when notation is case insensitive"() {
+        given:
+            project.enforce {
+                srcPath = SRC_PATH
+            }
+
+            File testClass1 = new File(Paths.get(SRC_PATH, "test", "Class1Test.cls").toString())
+            testClass1.write("@IsTest\n" +
+                    "public class Class1Test {}")
+
+            File testClass2 = new File(Paths.get(SRC_PATH, "test", "Class2Test.cls").toString())
+            testClass2.write("@ISTEST\n" +
+                    "public class Class2Test {}")
+
+            File testClass3 = new File(Paths.get(SRC_PATH, "test", "Class3Test.cls").toString())
+            testClass3.write("@istest\n" +
+                    "public class Class3Test {}")
+
+            RunTestTask runTestTask = project.tasks.findByName(RUN_TEST_TASK_NAME) as RunTestTask
+            String path = Paths.get(SRC_PATH, "test")
+        when:
+            ArrayList<String> classNames = runTestTask.getClassNames(path, "*Test*")
+        then:
+            classNames.sort() == ["FGW_Console_CTRLTest", "FGW_APIFactoryTest", "Class1Test", "Class2Test",
+                                  "Class3Test"].sort()
+    }
+
+    def cleanupSpec() {
+        new File(Paths.get(SRC_PATH, "test", "Class1Test.cls").toString()).delete()
+        new File(Paths.get(SRC_PATH, "test", "Class2Test.cls").toString()).delete()
+        new File(Paths.get(SRC_PATH, "test", "Class3Test.cls").toString()).delete()
     }
 }